feat(memory): reorganize controllers and remove deprecated endpoints
- Add new memory_controller.py and memory_analytics_controller.py modules
- Register new memory and analytics routers in manager_router
- Remove deprecated /writer_service synchronous write endpoint from memory_agent_controller
- Remove deprecated /write asynchronous write endpoint and related Celery task handling
- Refactor memory_agent_controller to remove unused imports (html, Header, InterestMemoryCache, etc.)
- Simplify memory_agent_controller dependencies and remove workspace access guards
- Consolidate write operations to dedicated memory controller for cleaner separation of concerns
