+++ api/app/controllers/memory_agent_controller.py
@@ -1,32 +1,23 @@
-import html
 from typing import List, Optional
 
 from dotenv import load_dotenv
-from fastapi import APIRouter, Depends, File, Form, Query, UploadFile, Header
+from fastapi import APIRouter, Depends, File, Form, Query, UploadFile
 from sqlalchemy.orm import Session
 from starlette.responses import StreamingResponse
 
-from app.cache.memory.interest_memory import InterestMemoryCache
-from app.celery_app import celery_app
 from app.core.error_codes import BizCode
-from app.core.language_utils import get_language_from_header
 from app.core.logging_config import get_api_logger
-from app.core.memory.enums import SearchStrategy, Neo4jNodeType
-from app.core.memory.memory_service import MemoryService
 from app.core.rag.llm.cv_model import QWenCV
 from app.core.response_utils import fail, success
 from app.db import get_db
-from app.dependencies import cur_workspace_access_guard, get_current_user
+from app.dependencies import get_current_user
 from app.models import ModelApiKey
 from app.models.user_model import User
-from app.repositories import knowledge_repository
-from app.schemas.memory_agent_schema import StorageType, UserInput, Write_UserInput, WriteMemoryRequest
+from app.schemas.memory_agent_schema import Write_UserInput
 from app.schemas.response_schema import ApiResponse
-from app.services import task_service, workspace_service
 from app.services.memory_agent_service import MemoryAgentService
 from app.services.memory_agent_service import get_end_user_connected_config as get_config
 from app.services.model_service import ModelConfigService
-from app.utils.tmp_session import ChatSessionCache
 
 load_dotenv()
 api_logger = get_api_logger()
@@ -121,300 +112,6 @@ async def download_log(
             return fail(BizCode.INTERNAL_ERROR, "启动日志流式传输失败", str(e))
 
 
-# [DEPRECATED] 下一版本移除：/writer_service 同步写入端点将改为统一走 dispatcher → push_task 异步路径
-@router.post("/writer_service", response_model=ApiResponse)
-@cur_workspace_access_guard()
-async def write_server(
-        user_input: Write_UserInput,
-        language_type: str = Header(default=None, alias="X-Language-Type"),
-        db: Session = Depends(get_db),
-        current_user: User = Depends(get_current_user)
-):
-    """
-    Write service endpoint - processes write operations synchronously
-
-    Args:
-        user_input: Write request containing message and end_user_id
-        language_type: 语言类型 ("zh" 中文, "en" 英文)，通过 X-Language-Type Header 传递
-
-    Returns:
-        Response with write operation status
-    """
-    # 使用集中化的语言校验
-    language = get_language_from_header(language_type)
-
-    config_id = user_input.config_id
-    workspace_id = current_user.current_workspace_id
-    api_logger.info(f"Write service: workspace_id={workspace_id}, config_id={config_id}, language_type={language}")
-
-    # 获取 storage_type，如果为 None 则使用默认值
-    storage_type = workspace_service.get_workspace_storage_type(
-        db=db,
-        workspace_id=workspace_id,
-        user=current_user
-    )
-    if storage_type is None: storage_type = 'neo4j'
-    user_rag_memory_id = ''
-
-    # 如果 storage_type 是 rag，必须确保有有效的 user_rag_memory_id
-    if storage_type == 'rag':
-        if workspace_id:
-            knowledge = knowledge_repository.get_knowledge_by_name(
-                db=db,
-                name="USER_RAG_MERORY",
-                workspace_id=workspace_id
-            )
-            if knowledge:
-                user_rag_memory_id = str(knowledge.id)
-            else:
-                api_logger.warning(
-                    f"未找到名为 'USER_RAG_MERORY' 的知识库，workspace_id: {workspace_id}，将使用 neo4j 存储")
-                storage_type = 'neo4j'
-        else:
-            api_logger.warning("workspace_id 为空，无法使用 rag 存储，将使用 neo4j 存储")
-            storage_type = 'neo4j'
-
-    api_logger.info(
-        f"Write service requested for group {user_input.end_user_id}, storage_type: {storage_type}, user_rag_memory_id: {user_rag_memory_id}")
-    try:
-        messages_list = memory_agent_service.get_messages_list(user_input)
-        result = await memory_agent_service.write_memory(
-            WriteMemoryRequest(
-                end_user_id=user_input.end_user_id,
-                messages=messages_list,
-                config_id=config_id,
-                storage_type=storage_type,
-                user_rag_memory_id=user_rag_memory_id,
-                language=language,
-                conversation_id=user_input.conversation_id,
-            ),
-            db,
-        )
-
-        return success(data=result, msg="写入成功")
-    except BaseException as e:
-        # Handle ExceptionGroup from TaskGroup (Python 3.11+) or BaseExceptionGroup
-        if hasattr(e, 'exceptions'):
-            error_messages = [f"{type(sub_e).__name__}: {str(sub_e)}" for sub_e in e.exceptions]
-            detailed_error = "; ".join(error_messages)
-            api_logger.error(f"Write operation error (TaskGroup): {detailed_error}", exc_info=True)
-            return fail(BizCode.INTERNAL_ERROR, "写入失败", detailed_error)
-        api_logger.error(f"Write operation error: {str(e)}", exc_info=True)
-        return fail(BizCode.INTERNAL_ERROR, "写入失败", str(e))
-
-
-@router.post("/write", response_model=ApiResponse)
-@cur_workspace_access_guard()
-async def write_server_async(
-        user_input: Write_UserInput,
-        language_type: str = Header(default=None, alias="X-Language-Type"),
-        db: Session = Depends(get_db),
-        current_user: User = Depends(get_current_user)
-):
-    """
-    Async write service endpoint - enqueues write processing to Celery
-
-    Args:
-        user_input: Write request containing message and end_user_id
-        language_type: 语言类型 ("zh" 中文, "en" 英文)，通过 X-Language-Type Header 传递
-
-    Returns:
-        Task ID for tracking async operation
-        Use GET /memory/write_result/{task_id} to check task status and get result
-    """
-    # 使用集中化的语言校验
-    language = get_language_from_header(language_type)
-
-    config_id = user_input.config_id
-    workspace_id = current_user.current_workspace_id
-    api_logger.info(
-        f"Async write service: workspace_id={workspace_id}, config_id={config_id}, language_type={language}")
-
-    # 获取 storage_type，如果为 None 则使用默认值
-    storage_type = workspace_service.get_workspace_storage_type(
-        db=db,
-        workspace_id=workspace_id,
-        user=current_user
-    )
-    if storage_type is None: storage_type = 'neo4j'
-    user_rag_memory_id = ''
-    if workspace_id:
-
-        knowledge = knowledge_repository.get_knowledge_by_name(
-            db=db,
-            name="USER_RAG_MERORY",
-            workspace_id=workspace_id
-        )
-        if knowledge: user_rag_memory_id = str(knowledge.id)
-    api_logger.info(f"Async write: storage_type={storage_type}, user_rag_memory_id={user_rag_memory_id}")
-    try:
-        # ── RAG 路径：保持不变，直接拼接文本写向量库 ──
-        if storage_type and storage_type.lower() == StorageType.RAG.value:
-            from app.core.memory.memory_service import MemoryService
-            messages_list = memory_agent_service.get_messages_list(user_input)
-            await MemoryService.write_messages_to_rag(
-                messages=messages_list,
-                end_user_id=user_input.end_user_id,
-                user_rag_memory_id=user_rag_memory_id,
-            )
-            api_logger.info(f"RAG write completed for end_user={user_input.end_user_id}")
-            return success(data={}, msg="RAG 写入完成")
-
-        # ── Neo4j 路径：通过 dispatcher 写入 ──
-        from app.core.memory.memory_service import MemoryService as _MemoryService
-
-        workspace_id_str = str(current_user.current_workspace_id) if current_user.current_workspace_id else ""
-        messages_list = memory_agent_service.get_messages_list(user_input)
-
-        task_ids = await _MemoryService.dispatch_api_service_async(
-            messages=messages_list,
-            end_user_id=user_input.end_user_id,
-            config_id=config_id,
-            workspace_id=workspace_id_str,
-            language=language,
-        )
-
-        api_logger.info(
-            f"Write tasks queued: {len(task_ids)} tasks, end_user={user_input.end_user_id}"
-        )
-
-        return success(data={"task_ids": task_ids}, msg=f"已提交 {len(task_ids)} 个写入任务")
-    except Exception as e:
-        api_logger.error(f"Async write operation failed: {str(e)}")
-        return fail(BizCode.INTERNAL_ERROR, "写入失败", str(e))
-
-
-@router.post("/read_service", response_model=ApiResponse)
-@cur_workspace_access_guard()
-async def read_server(
-        user_input: UserInput,
-        db: Session = Depends(get_db),
-        current_user: User = Depends(get_current_user)
-):
-    """
-    Read service endpoint - processes read operations synchronously
-    
-    search_switch values:
-    - "0": Requires verification
-    - "1": No verification, direct split
-    - "2": Direct answer based on context
-    
-    Args:
-        user_input: Read request with message, history, search_switch, and end_user_id
-    
-    Returns:
-        Response with query answer
-    """
-    config_id = user_input.config_id
-    workspace_id = current_user.current_workspace_id
-
-    storage_type = workspace_service.get_workspace_storage_type(
-        db=db,
-        workspace_id=workspace_id,
-        user=current_user
-    )
-    if storage_type is None: storage_type = 'neo4j'
-    user_rag_memory_id = ''
-    if workspace_id:
-
-        knowledge = knowledge_repository.get_knowledge_by_name(
-            db=db,
-            name="USER_RAG_MERORY",
-            workspace_id=workspace_id
-        )
-        if knowledge:
-            user_rag_memory_id = str(knowledge.id)
-
-    session_id = user_input.session_id.hex
-
-    api_logger.info(
-        f"Read service: group={user_input.end_user_id}, storage_type={storage_type}, user_rag_memory_id={user_rag_memory_id}, workspace_id={workspace_id}, session_id={session_id}")
-    try:
-        memory_config = get_config(user_input.end_user_id, db)
-        service = MemoryService(
-            memory_config["memory_config_id"],
-            end_user_id=user_input.end_user_id,
-            draft=True
-        )
-        session_cache = ChatSessionCache(session_id)
-        search_result = await service.read(
-            user_input.message,
-            SearchStrategy(user_input.search_switch),
-            history=await session_cache.get_history(),
-        )
-        intermediate_outputs = []
-        sub_queries = set()
-        for memory in search_result.memories:
-            sub_queries.add(str(memory.query))
-        idx = 0
-        if user_input.search_switch in [SearchStrategy.DEEP, SearchStrategy.NORMAL]:
-            intermediate_outputs.append({
-                "type": "problem_split",
-                "title": "问题拆分",
-                "data": [
-                    {
-                        "id": f"Q{(idx := idx + 1)}",
-                        "question": question
-                    }
-                    for question in sub_queries
-                    if question
-                ]
-            })
-        perceptual_data = [
-            memory.data
-            for memory in search_result.memories
-            if memory.source == Neo4jNodeType.PERCEPTUAL
-        ]
-
-        if len(perceptual_data):
-            intermediate_outputs.append({
-                "type": "perceptual_retrieve",
-                "title": "感知记忆检索",
-                "data": perceptual_data,
-                "total": len(perceptual_data),
-            })
-        intermediate_outputs.append({
-            "type": "search_result",
-            "title": f"合并检索结果 (共{len(sub_queries)}个查询,{len(search_result.memories)}条结果)",
-            "result": search_result.content,
-            "raw_result": search_result.memories,
-            "total": len(search_result.memories),
-        })
-        answer = await memory_agent_service.generate_summary_from_retrieve(
-            end_user_id=user_input.end_user_id,
-            retrieve_info=search_result.content,
-            history=[],
-            query=user_input.message,
-            config_id=config_id,
-            db=db
-        )
-        await session_cache.append_many(
-            [
-                {"role": "user", "content": user_input.message},
-                {"role": "assistant", "content": answer}
-            ]
-        )
-        for _ in intermediate_outputs:
-            if _["type"] == "search_result":
-                _["result"] = html.escape(_["result"])
-        result = {
-            'answer': answer,
-            "intermediate_outputs": intermediate_outputs,
-            "session_id": session_id,
-        }
-
-        return success(data=result, msg="回复对话消息成功")
-    except BaseException as e:
-        # Handle ExceptionGroup from TaskGroup (Python 3.11+) or BaseExceptionGroup
-        if hasattr(e, 'exceptions'):
-            error_messages = [f"{type(sub_e).__name__}: {str(sub_e)}" for sub_e in e.exceptions]
-            detailed_error = "; ".join(error_messages)
-            api_logger.error(f"Read operation error (TaskGroup): {detailed_error}", exc_info=True)
-            return fail(BizCode.INTERNAL_ERROR, "回复对话消息失败", detailed_error)
-        api_logger.error(f"Read operation error: {str(e)}", exc_info=True)
-        return fail(BizCode.INTERNAL_ERROR, "回复对话消息失败", str(e))
-
-
 @router.post("/file", response_model=ApiResponse)
 async def file_update(
         files: List[UploadFile] = File(..., description="要上传的文件"),
@@ -465,205 +162,6 @@ async def file_update(
     except Exception as e:
         api_logger.error(f"File processing failed: {str(e)}", exc_info=True)
         return fail(BizCode.INTERNAL_ERROR, "转换文本失败", str(e))
-
-
-@router.post("/read_service_async", response_model=ApiResponse)
-@cur_workspace_access_guard()
-async def read_server_async(
-        user_input: UserInput,
-        db: Session = Depends(get_db),
-        current_user: User = Depends(get_current_user)
-):
-    config_id = user_input.config_id
-    workspace_id = current_user.current_workspace_id
-    api_logger.info(f"Async read service: workspace_id={workspace_id}, config_id={config_id}")
-
-    # 获取 storage_type，如果为 None 则使用默认值
-    storage_type = workspace_service.get_workspace_storage_type(
-        db=db,
-        workspace_id=workspace_id,
-        user=current_user
-    )
-    if storage_type is None: storage_type = 'neo4j'
-    user_rag_memory_id = ''
-    if workspace_id:
-
-        knowledge = knowledge_repository.get_knowledge_by_name(
-            db=db,
-            name="USER_RAG_MERORY",
-            workspace_id=workspace_id
-        )
-        if knowledge: user_rag_memory_id = str(knowledge.id)
-    api_logger.info(f"Async read: storage_type={storage_type}, user_rag_memory_id={user_rag_memory_id}")
-    try:
-        session_id = user_input.session_id.hex
-        session_cache = ChatSessionCache(session_id)
-        task = celery_app.send_task(
-            "app.core.memory.agent.read_message",
-            args=[user_input.end_user_id, user_input.message, await session_cache.get_history(),
-                  user_input.search_switch,
-                  config_id, storage_type, user_rag_memory_id]
-        )
-        api_logger.info(f"Read task queued: {task.id}")
-
-        return success(data={"task_id": task.id}, msg="查询任务已提交")
-    except Exception as e:
-        api_logger.error(f"Async read operation failed: {str(e)}")
-        return fail(BizCode.INTERNAL_ERROR, "回复对话消息失败", str(e))
-
-
-@router.get("/read_result/", response_model=ApiResponse)
-async def get_read_task_result(
-        task_id: str,
-        current_user: User = Depends(get_current_user)
-):
-    """
-    Get the status and result of an async read task
-    
-    Args:
-        task_id: Celery task ID returned from /read_service_async
-    
-    Returns:
-        Task status and result if completed
-        
-    Response format:
-        - PENDING: Task is waiting to be executed
-        - STARTED: Task has started
-        - SUCCESS: Task completed successfully, returns result
-        - FAILURE: Task failed, returns error message
-    """
-    api_logger.info(f"Read task status check requested for task {task_id}")
-    try:
-        result = task_service.get_task_memory_read_result(task_id)
-        status = result.get("status")
-
-        if status == "SUCCESS":
-            # 任务成功完成
-            task_result = result.get("result", {})
-            if isinstance(task_result, dict):
-                # 新格式：包含详细信息
-                return success(
-                    data={
-                        "result": task_result.get("result"),
-                        "end_user_id": task_result.get("end_user_id"),
-                        "elapsed_time": task_result.get("elapsed_time"),
-                        "task_id": task_id
-                    },
-                    msg="查询任务已完成"
-                )
-            else:
-                # 旧格式：直接返回结果
-                return success(data=task_result, msg="查询任务已完成")
-
-        elif status == "FAILURE":
-            # 任务失败
-            error_info = result.get("result", "Unknown error")
-            if isinstance(error_info, dict):
-                error_msg = error_info.get("error", str(error_info))
-            else:
-                error_msg = str(error_info)
-            return fail(BizCode.INTERNAL_ERROR, "查询任务失败", error_msg)
-
-        elif status in ["PENDING", "STARTED"]:
-            # 任务进行中
-            return success(
-                data={
-                    "status": status,
-                    "task_id": task_id,
-                    "message": "任务处理中，请稍后查询"
-                },
-                msg="查询任务处理中"
-            )
-        else:
-            # 未知状态
-            return success(
-                data={
-                    "status": status,
-                    "task_id": task_id
-                },
-                msg=f"任务状态: {status}"
-            )
-
-    except Exception as e:
-        api_logger.error(f"Read task status check failed: {str(e)}", exc_info=True)
-        return fail(BizCode.INTERNAL_ERROR, "任务状态查询失败", str(e))
-
-
-@router.get("/write_result/", response_model=ApiResponse)
-async def get_write_task_result(
-        task_id: str,
-        current_user: User = Depends(get_current_user)
-):
-    """
-    Get the status and result of an async write task
-    
-    Args:
-        task_id: Celery task ID returned from /writer_service_async
-    
-    Returns:
-        Task status and result if completed
-        
-    Response format:
-        - PENDING: Task is waiting to be executed
-        - STARTED: Task has started
-        - SUCCESS: Task completed successfully, returns result
-        - FAILURE: Task failed, returns error message
-    """
-    api_logger.info(f"Write task status check requested for task {task_id}")
-    try:
-        result = task_service.get_task_memory_write_result(task_id)
-        status = result.get("status")
-
-        if status == "SUCCESS":
-            # 任务成功完成
-            task_result = result.get("result", {})
-            if isinstance(task_result, dict):
-                # 新格式：包含详细信息
-                return success(
-                    data={
-                        "result": task_result.get("result"),
-                        "end_user_id": task_result.get("end_user_id"),
-                        "elapsed_time": task_result.get("elapsed_time"),
-                        "task_id": task_id
-                    },
-                    msg="写入任务已完成"
-                )
-            else:
-                # 旧格式：直接返回结果
-                return success(data=task_result, msg="写入任务已完成")
-
-        elif status == "FAILURE":
-            # 任务失败
-            error_info = result.get("result", "Unknown error")
-            if isinstance(error_info, dict):
-                error_msg = error_info.get("error", str(error_info))
-            else:
-                error_msg = str(error_info)
-            return fail(BizCode.INTERNAL_ERROR, "写入任务失败", error_msg)
-
-        elif status in ["PENDING", "STARTED"]:
-            # 任务进行中
-            return success(
-                data={
-                    "status": status,
-                    "task_id": task_id,
-                    "message": "任务处理中，请稍后查询"
-                },
-                msg="写入任务处理中"
-            )
-        else:
-            # 未知状态
-            return success(
-                data={
-                    "status": status,
-                    "task_id": task_id
-                },
-                msg=f"任务状态: {status}"
-            )
-
-    except Exception as e:
-        api_logger.error(f"Write task status check failed: {str(e)}", exc_info=True)
-        return fail(BizCode.INTERNAL_ERROR, "任务状态查询失败", str(e))
 
 
 @router.post("/status_type", response_model=ApiResponse)
@@ -740,58 +238,6 @@ async def get_knowledge_type_stats_api(
     except Exception as e:
         api_logger.error(f"Knowledge type stats failed: {str(e)}")
         return fail(BizCode.INTERNAL_ERROR, "获取知识库类型统计失败", str(e))
-
-
-@router.get("/analytics/interest_distribution/by_user", response_model=ApiResponse)
-async def get_interest_distribution_by_user_api(
-        end_user_id: str = Query(..., description="用户ID（必填）"),
-        limit: int = Query(5, le=5, description="返回兴趣标签数量限制，最多5个"),
-        language_type: str = Header(default=None, alias="X-Language-Type"),
-        current_user: User = Depends(get_current_user),
-        db: Session = Depends(get_db),
-):
-    """
-    获取指定用户的兴趣分布标签
-    
-    与热门标签不同，此接口专注于识别用户的兴趣活动（运动、爱好、学习、创作等），
-    过滤掉纯物品、工具、地点等不代表用户主动参与活动的名词。
-    
-    返回格式：
-    [
-        {"name": "兴趣活动名", "frequency": 频次},
-        ...
-    ]
-    """
-    language = get_language_from_header(language_type)
-    api_logger.info(f"Interest distribution by user requested: end_user_id={end_user_id}, language={language}")
-    try:
-        # 优先读取缓存
-        cached = await InterestMemoryCache.get_interest_distribution(
-            end_user_id=end_user_id,
-            language=language,
-        )
-        if cached is not None:
-            api_logger.info(f"Interest distribution cache hit: end_user_id={end_user_id}")
-            return success(data=cached, msg="获取兴趣分布标签成功")
-
-        # 缓存未命中，调用模型生成
-        result = await memory_agent_service.get_interest_distribution_by_user(
-            end_user_id=end_user_id,
-            limit=limit,
-            language=language
-        )
-
-        # 写入缓存，24小时过期
-        await InterestMemoryCache.set_interest_distribution(
-            end_user_id=end_user_id,
-            language=language,
-            data=result,
-        )
-
-        return success(data=result, msg="获取兴趣分布标签成功")
-    except Exception as e:
-        api_logger.error(f"Interest distribution by user failed: {str(e)}")
-        return fail(BizCode.INTERNAL_ERROR, "获取兴趣分布标签失败", str(e))
 
 
 @router.get("/analytics/user_profile", response_model=ApiResponse)
